modal open/close, passing in props
