+++ src/components/ModalRecipeListDetail.js
@@ -1,0 +1,62 @@
+import React, { useEffect, useState } from 'react';
+import axios from 'axios';
+
+import {
+  Container,
+  Col,
+  Row,
+  Card,
+  CardBody,
+  CardTitle,
+  Button,
+  Modal,
+  ModalHeader,
+  ModalBody,
+  ModalFooter
+} from 'reactstrap';
+
+const ModalRecipeListDetail = (props) => {
+  const [modal, setModal] = useState(false);
+  const toggle = () => setModal(!modal);
+
+  return (
+    <div>
+      <Button onClick={toggle}>View Recipe</Button>
+      <Modal isOpen={modal} toggle={toggle} scrollable='true' fullscreen>
+        <ModalHeader toggle={toggle}>{props.title}</ModalHeader>
+        <img
+          className="align-self-center"
+          alt={props.title}
+          src={props.image}
+          style={{
+            width: '25rem'
+          }}
+        />
+        <ModalBody>
+          Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do
+          eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
+          minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+          aliquip ex ea commodo consequat. Duis aute irure dolor in
+          reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
+          pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
+          culpa qui officia deserunt mollit anim id est laborum.  Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do
+          eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
+          minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+          aliquip ex ea commodo consequat. Duis aute irure dolor in
+          reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
+          pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
+          culpa qui officia deserunt mollit anim id est laborum.  Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do
+          eiusmod tempor incididunt ut labore et dolore magna aliqua.
+        </ModalBody>
+        <ModalFooter>
+  
+          <Button color="primary" onClick={toggle}>
+            Add to Collection
+          </Button>
+  
+        </ModalFooter>
+      </Modal>
+    </div>
+  );
+};
+export default ModalRecipeListDetail;
