+++ src/pages/RecipeListPage.js
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
-
 import {
   Container,
   Col,
@@ -14,6 +13,7 @@ import {
   ModalBody,
   ModalFooter
 } from 'reactstrap';
+import ModalRecipeListDetail from '../components/ModalRecipeListDetail';
 
 const RecipeListPage = (props) => {
   const [responseData, setResponseData] = useState([]);
@@ -37,9 +37,8 @@ const RecipeListPage = (props) => {
   }, []);
   console.log(responseData);
 
-// The Modal:
-  const [modal, setModal] = useState(false);
-  const toggle = () => setModal(!modal);
+
+  
 
   return (
     <>
@@ -70,35 +69,9 @@ const RecipeListPage = (props) => {
                         {responseData.title}
                       </CardTitle>
                     </CardBody>
-                    <Button onClick={toggle}>View Recipe</Button>
-
-                    <Modal isOpen={modal} toggle={toggle} scrollable='true' fullscreen>
-                      <ModalHeader toggle={toggle}>Modal title</ModalHeader>
-                      <ModalBody>
-                        Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do
-                        eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
-                        minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-                        aliquip ex ea commodo consequat. Duis aute irure dolor in
-                        reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-                        pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-                        culpa qui officia deserunt mollit anim id est laborum.  Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do
-                        eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
-                        minim veniam, quis nostrud exercitation ullamco laboris nisi ut
-                        aliquip ex ea commodo consequat. Duis aute irure dolor in
-                        reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-                        pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-                        culpa qui officia deserunt mollit anim id est laborum.  Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do
-                        eiusmod tempor incididunt ut labore et dolore magna aliqua. 
-                      </ModalBody>
-                      <ModalFooter>
-                        <Button color="primary" onClick={toggle}>
-                          Do Something
-                        </Button>{' '}
-                        <Button color="secondary" onClick={toggle}>
-                          Cancel
-                        </Button>
-                      </ModalFooter>
-                    </Modal>
+                    
+                    <ModalRecipeListDetail detail={responseData} />
+                    
                   </Card>
                 </Col>
               )
